Fix: sink.text.pretty: assign status at the end of pretty_consume
The compiler on SLES 12 reports:

pretty.c:200:2: warning: 'status' may be used uninitialized in this function [-Wmaybe-uninitialized]
  return status;
  ^

We are indeed missing an assignment of status at the end, if everything
went well.  In practice, status has the right value (_OK) from the last
call of handle_message.  But if, for some reason,
bt_message_iterator_next returns OK with 0 messages, then it would stay
unset.  So assigning it explicitly at the end doesn't hurt.

Change-Id: Ic3bf018388f006c919173f9e8942a8ab2384c0f1
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/3427
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>